Move ExternalOperationException to GitExtUtils
