+++ GitUI/NBugReports/BugReporter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text;
 using System.Windows.Forms;
-using GitCommands;
+using GitExtUtils;
 using GitUI;
 using GitUI.NBugReports;
 using Microsoft.WindowsAPICodePack.Dialogs;
