+++ GitUI/NBugReports/UserExternalOperationException.cs
@@ -1,4 +1,4 @@
-﻿using GitCommands;
+﻿using GitExtUtils;
 
 namespace GitUI.NBugReports
 {
