+++ UnitTests/GitCommands.Tests/Git/IndexLockManagerTests.cs
@@ -5,6 +5,7 @@ using CommonTestUtils;
 using FluentAssertions;
 using GitCommands;
 using GitCommands.Git;
+using GitExtUtils;
 using GitUIPluginInterfaces;
 using NSubstitute;
 using NUnit.Framework;
